requirements.txt
apache-airflow-providers-amazon==2.2.0
boto3
